+++ controllers/virtualmachinesnapshot/virtualmachinesnapshot_controller_intg_test.go
@@ -145,6 +145,14 @@ func intgTestsReconcile() {
 					vm.Status.CurrentSnapshot = newLocalObjectRefWithSnapshotName(vmSnapshot.Name)
 					Expect(vcSimCtx.Client.Status().Update(ctx, vm)).To(Succeed())
 				}
+
+				provider.Lock()
+				provider.SyncVMSnapshotTreeStatusFn = func(_ context.Context, vm *vmopv1.VirtualMachine) error {
+					vm.Status.CurrentSnapshot = nil
+					vm.Status.RootSnapshots = []vmopv1common.LocalObjectRef{}
+					return nil
+				}
+				provider.Unlock()
 			})
 
 			JustBeforeEach(func() {
@@ -233,6 +241,32 @@ func intgTestsReconcile() {
 							g.Expect(vmObj).ToNot(BeNil())
 							g.Expect(vmObj.Status.CurrentSnapshot).To(Not(BeNil()))
 							g.Expect(*vmObj.Status.CurrentSnapshot).To(Equal(*newLocalObjectRefWithSnapshotName(vmSnapshot.Name)))
+						}).Should(Succeed())
+					})
+				})
+			})
+
+			When("Calling SyncVMSnapshotTreeStatus to VC", func() {
+				When("VC error", func() {
+					BeforeEach(func() {
+						provider.Lock()
+						provider.SyncVMSnapshotTreeStatusFn = func(_ context.Context, _ *vmopv1.VirtualMachine) error {
+							return errors.New("fubar")
+						}
+						provider.Unlock()
+					})
+					It("returns error, and current and root snapshots are not updated", func() {
+						vmSnapshotObjKey := types.NamespacedName{Name: vmSnapshot.Name, Namespace: vmSnapshot.Namespace}
+						Consistently(func(g Gomega) {
+							tmpVMSSnapshot := getVirtualMachineSnapshot(vcSimCtx, vmSnapshotObjKey)
+							g.Expect(tmpVMSSnapshot).To(Not(BeNil()))
+						}).Should(Succeed())
+						vmObjKey := types.NamespacedName{Name: vm.Name, Namespace: vm.Namespace}
+						Consistently(func(g Gomega) {
+							vmObj := getVirtualMachine(vcSimCtx, vmObjKey)
+							g.Expect(vmObj).ToNot(BeNil())
+							g.Expect(vmObj.Status.CurrentSnapshot).To(Equal(newLocalObjectRefWithSnapshotName(vmSnapshot.Name)))
+							g.Expect(vmObj.Status.RootSnapshots).To(BeNil())
 						}).Should(Succeed())
 					})
 				})
@@ -408,6 +442,11 @@ func intgTestsReconcile() {
 							Name: vmSnapshotL1.Name,
 						}, nil
 					}
+					provider.SyncVMSnapshotTreeStatusFn = func(_ context.Context, vm *vmopv1.VirtualMachine) error {
+						vm.Status.CurrentSnapshot = newLocalObjectRefWithSnapshotName(vmSnapshotL1.Name)
+						vm.Status.RootSnapshots = []vmopv1common.LocalObjectRef{*newLocalObjectRefWithSnapshotName(vmSnapshotL1.Name)}
+						return nil
+					}
 				})
 
 				JustBeforeEach(func() {
@@ -451,21 +490,6 @@ func intgTestsReconcile() {
 						}).Should(Succeed(), "waiting for vm root snapshots to be updated")
 					})
 				})
-
-				When("it's not the current snapshot", func() {
-					BeforeEach(func() {
-						currentSnapshotName = vmSnapshotL1Name
-					})
-
-					It("returns success, vm current snapshot is not changed", func() {
-						vmObjKey := types.NamespacedName{Name: vm.Name, Namespace: vm.Namespace}
-						Eventually(func(g Gomega) {
-							vmObj := getVirtualMachine(vcSimCtx, vmObjKey)
-							g.Expect(vmObj).ToNot(BeNil())
-							g.Expect(vmObj.Status.CurrentSnapshot.Name).To(Equal(vmSnapshotL1Name))
-						}).Should(Succeed())
-					})
-				})
 			})
 			When("root snapshot is deleted", func() {
 				//        L1  <--- current snapshot, deleted
@@ -479,6 +503,13 @@ func intgTestsReconcile() {
 				//       /   \
 				//   L3-n1    L3-n2
 
+				BeforeEach(func() {
+					provider.SyncVMSnapshotTreeStatusFn = func(_ context.Context, vm *vmopv1.VirtualMachine) error {
+						vm.Status.CurrentSnapshot = nil
+						vm.Status.RootSnapshots = []vmopv1common.LocalObjectRef{*newLocalObjectRefWithSnapshotName(vmSnapshotL2.Name)}
+						return nil
+					}
+				})
 				JustBeforeEach(func() {
 					By("delete the snapshot")
 					Expect(vcSimCtx.Client.Delete(ctx, vmSnapshotL1)).To(Succeed())
@@ -503,20 +534,6 @@ func intgTestsReconcile() {
 						}).Should(Succeed(), "waiting for vm root snapshots to be updated")
 					})
 				})
-
-				When("it's not the current snapshot", func() {
-					BeforeEach(func() {
-						currentSnapshotName = vmSnapshotL2Name
-					})
-					It("returns success, vm current snapshot is not changed", func() {
-						vmObjKey := types.NamespacedName{Name: vm.Name, Namespace: vm.Namespace}
-						Eventually(func(g Gomega) {
-							vmObj := getVirtualMachine(vcSimCtx, vmObjKey)
-							g.Expect(vmObj).ToNot(BeNil())
-							g.Expect(vmObj.Status.CurrentSnapshot.Name).To(Equal(vmSnapshotL2Name))
-						}).Should(Succeed())
-					})
-				})
 			})
 
 			When("leaf snapshot is deleted", func() {
@@ -538,6 +555,11 @@ func intgTestsReconcile() {
 						return &vimtypes.VirtualMachineSnapshotTree{
 							Name: vmSnapshotL2.Name,
 						}, nil
+					}
+					provider.SyncVMSnapshotTreeStatusFn = func(_ context.Context, vm *vmopv1.VirtualMachine) error {
+						vm.Status.CurrentSnapshot = newLocalObjectRefWithSnapshotName(vmSnapshotL2Name)
+						vm.Status.RootSnapshots = []vmopv1common.LocalObjectRef{*newLocalObjectRefWithSnapshotName(vmSnapshotL1.Name)}
+						return nil
 					}
 				})
 
@@ -574,20 +596,6 @@ func intgTestsReconcile() {
 							g.Expect(vmObj.Status.RootSnapshots).To(HaveLen(1))
 							g.Expect(vmObj.Status.RootSnapshots).To(ContainElement(*newLocalObjectRefWithSnapshotName(vmSnapshotL1.Name)))
 						}).Should(Succeed(), "waiting for vm root snapshots to be updated")
-					})
-				})
-
-				When("it's not the current snapshot", func() {
-					BeforeEach(func() {
-						currentSnapshotName = vmSnapshotL2Name
-					})
-					It("returns success, vm current snapshot is not changed", func() {
-						vmObjKey := types.NamespacedName{Name: vm.Name, Namespace: vm.Namespace}
-						Eventually(func(g Gomega) {
-							vmObj := getVirtualMachine(vcSimCtx, vmObjKey)
-							g.Expect(vmObj).ToNot(BeNil())
-							g.Expect(vmObj.Status.CurrentSnapshot.Name).To(Equal(vmSnapshotL2Name))
-						}).Should(Succeed())
 					})
 				})
 			})
