+++ pkg/providers/vm_provider_interface.go
@@ -79,4 +79,7 @@ type VirtualMachineProviderInterface interface {
 	GetParentSnapshot(ctx context.Context, vmSnapshotName string, vm *vmopv1.VirtualMachine) (*vimtypes.VirtualMachineSnapshotTree, error)
 	// SnapshotSize returns the size of a snapshot.
 	GetSnapshotSize(ctx context.Context, vmSnapshotName string, vm *vmopv1.VirtualMachine) (int64, error)
+
+	// SyncVMSnapshotTreeStatus syncs the VM's current and root snapshots status.
+	SyncVMSnapshotTreeStatus(ctx context.Context, vm *vmopv1.VirtualMachine) error
 }
