+++ pkg/providers/vsphere/vmprovider.go
@@ -7,7 +7,6 @@ package vsphere
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"math/rand"
 	"strings"
@@ -43,7 +42,6 @@ import (
 	"github.com/vmware-tanzu/vm-operator/pkg/providers/vsphere/contentlibrary"
 	vccreds "github.com/vmware-tanzu/vm-operator/pkg/providers/vsphere/credentials"
 	"github.com/vmware-tanzu/vm-operator/pkg/providers/vsphere/vcenter"
-	"github.com/vmware-tanzu/vm-operator/pkg/providers/vsphere/virtualmachine"
 	"github.com/vmware-tanzu/vm-operator/pkg/record"
 	"github.com/vmware-tanzu/vm-operator/pkg/topology"
 	pkgutil "github.com/vmware-tanzu/vm-operator/pkg/util"
@@ -629,122 +627,4 @@ func (vs *vSphereVMProvider) VSphereClient(
 		return nil, err
 	}
 	return c.Client, nil
-}
-
-// DeleteSnapshot deletes the snapshot from the VM.
-// The boolean indicating if the VM associated is deleted.
-func (vs *vSphereVMProvider) DeleteSnapshot(
-	ctx context.Context,
-	vmSnapshot *vmopv1.VirtualMachineSnapshot,
-	vm *vmopv1.VirtualMachine,
-	removeChildren bool,
-	consolidate *bool) (bool, error) {
-
-	logger := pkgutil.FromContextOrDefault(ctx).WithValues("vmName", vm.NamespacedName())
-	ctx = logr.NewContext(ctx, logger)
-
-	vmCtx := pkgctx.VirtualMachineContext{
-		Context: context.WithValue(ctx, vimtypes.ID{}, vs.getOpID(ctx, vm, "deleteSnapshot")),
-		Logger:  logger,
-		VM:      vm,
-	}
-
-	client, err := vs.getVcClient(ctx)
-	if err != nil {
-		return false, err
-	}
-
-	vcVM, err := vs.getVM(vmCtx, client, false)
-	if err != nil {
-		return false, fmt.Errorf("failed to get VirtualMachine %q: %w", vmCtx.VM.Name, err)
-	} else if vcVM == nil {
-		return true, nil
-	}
-
-	if err := virtualmachine.DeleteSnapshot(virtualmachine.SnapshotArgs{
-		VMSnapshot:     *vmSnapshot,
-		VcVM:           vcVM,
-		VMCtx:          vmCtx,
-		RemoveChildren: removeChildren,
-		Consolidate:    consolidate,
-	}); err != nil {
-		if errors.Is(err, virtualmachine.ErrSnapshotNotFound) {
-			vmCtx.Logger.V(5).Info("snapshot not found")
-			return false, nil
-		}
-		return false, err
-	}
-
-	return false, nil
-}
-
-// GetSnapshotSize gets the size of the snapshot from the VM.
-func (vs *vSphereVMProvider) GetSnapshotSize(
-	ctx context.Context,
-	vmSnapshotName string,
-	vm *vmopv1.VirtualMachine) (int64, error) {
-
-	logger := pkgutil.FromContextOrDefault(ctx).WithValues("vmName", vm.NamespacedName())
-	ctx = logr.NewContext(ctx, logger)
-
-	vmCtx := pkgctx.VirtualMachineContext{
-		Context: context.WithValue(ctx, vimtypes.ID{}, vs.getOpID(ctx, vm, "getSnapshotSize")),
-		Logger:  logger,
-		VM:      vm,
-	}
-
-	client, err := vs.getVcClient(ctx)
-	if err != nil {
-		return 0, err
-	}
-
-	vcVM, err := vs.getVM(vmCtx, client, true)
-	if err != nil {
-		return 0, fmt.Errorf("failed to get VirtualMachine %q: %w", vmCtx.VM.Name, err)
-	}
-
-	var moVM mo.VirtualMachine
-
-	if err = vcVM.Properties(ctx, vcVM.Reference(), []string{"snapshot", "layoutEx", "config.hardware.device"}, &moVM); err != nil {
-		return 0, err
-	}
-
-	vmCtx.MoVM = moVM
-
-	vmNapshot, err := virtualmachine.FindSnapshot(vmCtx, vmSnapshotName)
-	if err != nil {
-		return 0, fmt.Errorf("failed to find snapshot %q: %w", vmSnapshotName, err)
-	}
-
-	size := virtualmachine.GetSnapshotSize(
-		vmCtx, vmNapshot)
-
-	return size, nil
-}
-
-func (vs *vSphereVMProvider) GetParentSnapshot(
-	ctx context.Context,
-	vmSnapshotName string,
-	vm *vmopv1.VirtualMachine) (*vimtypes.VirtualMachineSnapshotTree, error) {
-
-	logger := pkgutil.FromContextOrDefault(ctx).WithValues("vmName", vm.NamespacedName())
-	ctx = logr.NewContext(ctx, logger)
-
-	vmCtx := pkgctx.VirtualMachineContext{
-		Context: context.WithValue(ctx, vimtypes.ID{}, vs.getOpID(ctx, vm, "getParentSnapshot")),
-		Logger:  logger,
-		VM:      vm,
-	}
-
-	client, err := vs.getVcClient(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	vcVM, err := vs.getVM(vmCtx, client, true)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get VirtualMachine %q: %w", vmCtx.VM.Name, err)
-	}
-
-	return virtualmachine.GetParentSnapshot(vmCtx, vcVM, vmSnapshotName)
 }
