+++ pkg/providers/vsphere/vmprovider_vm.go
@@ -11,7 +11,6 @@ import (
 	"maps"
 	"math/rand"
 	"path"
-	"slices"
 	"sort"
 	"strings"
 	"sync"
@@ -1380,7 +1379,7 @@ func (vs *vSphereVMProvider) reconcileSnapshotRevert(
 		"afterRestoreLabels", vmCtx.VM.Labels,
 		"afterRestoreSpecCurrentSnapshot", vmCtx.VM.Spec.CurrentSnapshot)
 
-	if err := vmlifecycle.UpdateCurrentSnapshotStatus(vmCtx, vs.k8sClient); err != nil {
+	if err := vmlifecycle.SyncVMSnapshotTreeStatus(vmCtx, vs.k8sClient); err != nil {
 		return true, err
 	}
 
@@ -1645,23 +1644,17 @@ func (vs *vSphereVMProvider) reconcileSnapshot(
 		return err
 	}
 
+	// Update the VM's status first before marking the snapshot as ready.
+	// In case when the VM update fails, the next reconcile will skip
+	// reconciling this snapshot if it's marked as ready.
+	if err := vs.updateVMStatus(vmCtx, snapArgs, *snapshotToProcess); err != nil {
+		vmCtx.Logger.Error(err, "Failed to update VM status", "snapshotName", snapshotToProcess.Name)
+		return err
+	}
+
 	// Update the snapshot status with the successful result
 	if err = PatchSnapshotSuccessStatus(vmCtx, vs.k8sClient, snapshotToProcess, snapMoRef); err != nil {
 		vmCtx.Logger.Error(err, "Failed to update snapshot status", "snapshotName", snapshotToProcess.Name)
-		return err
-	}
-
-	// Update the children snapshot status of parent snapshot
-	if err = vs.patchChildrenSnapshotStatus(snapArgs); err != nil {
-		vmCtx.Logger.Error(err, "Failed to update children snapshot status in its parent", "childrenSnapshotName", snapshotToProcess.Name)
-		return err
-	}
-
-	if err := vmlifecycle.UpdateCurrentSnapshotStatus(vmCtx, vs.k8sClient); err != nil {
-		return err
-	}
-
-	if err := vmlifecycle.UpdateRootSnapshots(vmCtx, vs.k8sClient); err != nil {
 		return err
 	}
 
@@ -3009,119 +3002,4 @@ func (vs *vSphereVMProvider) vmResizeGetArgs(
 	}
 
 	return resizeArgs, nil
-}
-
-// getVirtualMachineSnapshotsForVM finds all VirtualMachineSnapshot objects that reference this VM.
-func (vs *vSphereVMProvider) getVirtualMachineSnapshotsForVM(
-	vmCtx pkgctx.VirtualMachineContext) ([]vmopv1.VirtualMachineSnapshot, error) {
-
-	// List all VirtualMachineSnapshot objects in the VM's namespace
-	var snapshotList vmopv1.VirtualMachineSnapshotList
-	if err := vs.k8sClient.List(vmCtx, &snapshotList, ctrlclient.InNamespace(vmCtx.VM.Namespace)); err != nil {
-		return nil, fmt.Errorf("failed to list VirtualMachineSnapshot objects: %w", err)
-	}
-
-	// Filter snapshots that reference this VM. We do this by checking
-	// the VMRef, and by filtering the snapshots owned by this VM.
-	var vmSnapshots []vmopv1.VirtualMachineSnapshot
-	for _, snapshot := range snapshotList.Items {
-		if snapshot.Spec.VMRef != nil && snapshot.Spec.VMRef.Name == vmCtx.VM.Name {
-			vmSnapshots = append(vmSnapshots, snapshot)
-		}
-	}
-
-	vmCtx.Logger.V(4).Info("Found VirtualMachineSnapshot objects for VM",
-		"count", len(vmSnapshots))
-
-	return vmSnapshots, nil
-}
-
-// markSnapshotInProgress marks a snapshot as currently being processed.
-func (vs *vSphereVMProvider) markSnapshotInProgress(vmCtx pkgctx.VirtualMachineContext, vmSnapshot *vmopv1.VirtualMachineSnapshot) error {
-	// Create a patch to set the InProgress condition
-	patch := ctrlclient.MergeFrom(vmSnapshot.DeepCopy())
-
-	// Set the InProgress condition to prevent concurrent operations
-	pkgcnd.MarkFalse(
-		vmSnapshot,
-		vmopv1.VirtualMachineSnapshotReadyCondition,
-		vmopv1.VirtualMachineSnapshotInProgressReason,
-		"snapshot in progress")
-
-	// Use Patch instead of Update to avoid conflicts with snapshot controller
-	if err := vs.k8sClient.Status().Patch(vmCtx, vmSnapshot, patch); err != nil {
-		return fmt.Errorf("failed to mark snapshot as in progress: %w", err)
-	}
-
-	vmCtx.Logger.V(4).Info("Marked snapshot as in progress", "snapshotName", vmSnapshot.Name)
-	return nil
-}
-
-// markSnapshotFailed marks a snapshot as failed and clears the in-progress status.
-func (vs *vSphereVMProvider) markSnapshotFailed(vmCtx pkgctx.VirtualMachineContext, vmSnapshot *vmopv1.VirtualMachineSnapshot, err error) {
-	// Create a patch to update the snapshot status
-	patch := ctrlclient.MergeFrom(vmSnapshot.DeepCopy())
-
-	// Mark the snapshot as failed
-	pkgcnd.MarkFalse(vmSnapshot, vmopv1.VirtualMachineSnapshotReadyCondition, "SnapshotFailed", "%s", err.Error())
-
-	// Use Patch instead of Update to avoid conflicts with snapshot controller (best effort)
-	if updateErr := vs.k8sClient.Status().Patch(vmCtx, vmSnapshot, patch); updateErr != nil {
-		vmCtx.Logger.Error(updateErr, "Failed to update snapshot status after failure", "snapshotName", vmSnapshot.Name)
-	}
-
-	vmCtx.Logger.V(4).Info("Marked snapshot as failed", "snapshotName", vmSnapshot.Name, "error", err.Error())
-}
-
-// patchChildrenSnapshotStatus find the parent snapshot of current snapshot
-// and patches the children snapshot status of the parent snapshot by adding
-// the current snapshot to the children list of the parent snapshot.
-func (vs *vSphereVMProvider) patchChildrenSnapshotStatus(snapArgs virtualmachine.SnapshotArgs) error {
-	// Use vsphereprovider's GetParentSnapshot since it will fetch latest
-	// version of moVM from vsphere, after a new snapshot was taken above.
-	vmCtx := snapArgs.VMCtx
-	vmSnapshot := snapArgs.VMSnapshot
-
-	parentSnapshot, err := virtualmachine.GetParentSnapshot(vmCtx, snapArgs.VcVM, vmSnapshot.Name)
-	if err != nil {
-		vmCtx.Logger.Error(err, "failed to get parent snapshot", "snapshotName", vmSnapshot.Name)
-		return err
-	}
-
-	if parentSnapshot == nil {
-		vmCtx.Logger.V(4).Info("No parent snapshot found for snapshot, skipping children snapshot status update",
-			"snapshotName", vmSnapshot.Name)
-		return nil
-	}
-
-	parentCR := &vmopv1.VirtualMachineSnapshot{}
-	if err := vs.k8sClient.Get(vmCtx,
-		ctrlclient.ObjectKey{Name: parentSnapshot.Name, Namespace: vmSnapshot.Namespace}, parentCR); err != nil {
-		if apierrors.IsNotFound(err) {
-			// TODO (lubron): The parent snapshot might be externally created (only exist in vSphere)
-			// In this case, we should just skip the update.
-			// For now let's return an error.
-			return fmt.Errorf("parent snapshot %s/%s not found", parentSnapshot.Name, vmSnapshot.Namespace)
-		}
-		return fmt.Errorf("failed to get parent snapshot %s/%s: err: %s", parentSnapshot.Name, vmSnapshot.Namespace, err.Error())
-	}
-
-	if slices.ContainsFunc(parentCR.Status.Children, func(c common.LocalObjectRef) bool {
-		return c.Name == vmSnapshot.Name
-	}) {
-		vmCtx.Logger.V(4).Info("Snapshot already in children list of parent, skipping",
-			"parentSnapshotName", parentSnapshot.Name,
-			"snapshotName", vmSnapshot.Name)
-		return nil
-	}
-
-	objPatch := ctrlclient.MergeFrom(parentCR.DeepCopy())
-
-	parentCR.Status.Children = append(parentCR.Status.Children, common.LocalObjectRef{
-		Name:       vmSnapshot.Name,
-		APIVersion: vmopv1.GroupVersion.String(),
-		Kind:       "VirtualMachineSnapshot",
-	})
-
-	return vs.k8sClient.Status().Patch(vmCtx, parentCR, objPatch)
 }
