+++ pkg/providers/vsphere/vsphere_suite_test.go
@@ -33,8 +33,8 @@ func vcSimTests() {
 	Describe("VirtualMachineE2E", vmE2ETests)
 	Describe("VirtualMachineResize", vmResizeTests)
 	Describe("VirtualMachineUtilsTest", vmUtilTests)
-	Describe("VirtualMachineSnapshot", vmSnapshotTests)
 	Describe("VirtualMachineGroup", vmGroupTests)
+	Describe("VirtualMachineSnapshot", vmSnapshotTests)
 }
 
 func TestVSphereProvider(t *testing.T) {
